+++ packages/dropdown_button2/lib/src/dropdown_style_data.dart
@@ -184,18 +184,10 @@ class DropdownStyleData {
 class MenuItemStyleData {
   /// Creates a MenuItemStyleData.
   const MenuItemStyleData({
-    this.height = _kMenuItemHeight,
-    this.customHeights,
     this.padding,
     this.overlayColor,
     this.selectedMenuItemBuilder,
   });
-
-  /// The height of the menu item, default value is [kMinInteractiveDimension]
-  final double height;
-
-  /// Define different heights for the menu items (useful for adding dividers)
-  final List<double>? customHeights;
 
   /// The padding of menu items
   final EdgeInsetsGeometry? padding;