+++ packages/dropdown_button2/lib/src/utils.dart
@@ -24,3 +24,12 @@ double _clampDouble(double x, double min, double max) {
   }
   return x;
 }
+
+// ignore: public_member_api_docs
+extension ExtendedIterable<E> on Iterable<E> {
+  /// Like Iterable<T>.map but the callback has index as second argument
+  Iterable<T> mapIndexed<T>(T Function(E e, int i) f) {
+    var i = 0;
+    return map((e) => f(e, i++));
+  }
+}