+++ packages/dropdown_button2_test/lib/src/few_styling_example.dart
@@ -54,6 +54,7 @@ class _FewStylingExampleState extends State<FewStylingExample> {
             items: items
                 .map((String item) => DropdownItem<String>(
                       value: item,
+                      height: 40,
                       child: Text(
                         item,
                         style: const TextStyle(
@@ -107,7 +108,6 @@ class _FewStylingExampleState extends State<FewStylingExample> {
               ),
             ),
             menuItemStyleData: const MenuItemStyleData(
-              height: 40,
               padding: EdgeInsets.only(left: 14, right: 14),
             ),
           ),